backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
marshmallow==3.20.1
email-validator==2.1.0
gunicorn==21.2.0
